refactor(arch): Phase 1 — infrastructure/{http,utils} + domain/{time,market,chart} → shared/{api,lib}
- 6개 파일 이동: isBot→shared/api, dateKey/withRetry→shared/lib, eastern/marketSession/timeFormat→shared/lib
- 32개 파일 import path 갱신
- 빈 디렉토리 5개 자동 정리 (infrastructure/http, infrastructure/utils, domain/time, domain/chart, domain/market)

Phase 1 of FSD migration (PR B: shared api + domain utils)
