+++ src/components/options/OptionsStaleDataBanner.tsx
@@ -5,7 +5,7 @@ import {
     KST_EDT_HOURS_DISPLAY,
     KST_EST_HOURS_DISPLAY,
 } from '@/lib/options/marketHoursDisplay';
-import { EDT_OFFSET_HOURS, getEasternOffsetHours } from '@/domain/time/eastern';
+import { EDT_OFFSET_HOURS, getEasternOffsetHours } from '@/shared/lib/eastern';
 
 /**
  * Surfaces a "data temporarily empty" notice when the upstream provider
