+++ src/infrastructure/db/agreementRepository.ts
@@ -1,7 +1,7 @@
 import { NEON_TRANSIENT_RETRY } from '@/infrastructure/db/isNeonTransientError';
 import { agreements } from '@/infrastructure/db/schema';
 import type { SiglensDatabase } from '@/infrastructure/db/types';
-import { withRetry } from '@/infrastructure/utils/withRetry';
+import { withRetry } from '@/shared/lib/withRetry';
 
 export interface AgreementInsertInput {
     userId: string;
