+++ src/infrastructure/db/contactRepository.ts
@@ -1,7 +1,7 @@
 import { NEON_TRANSIENT_RETRY } from '@/infrastructure/db/isNeonTransientError';
 import { inquiries } from '@/infrastructure/db/schema';
 import type { SiglensDatabase } from '@/infrastructure/db/types';
-import { withRetry } from '@/infrastructure/utils/withRetry';
+import { withRetry } from '@/shared/lib/withRetry';
 
 /** Input required to create a new inquiry record. */
 export interface ContactInput {
