+++ src/infrastructure/db/userApiKeyRepository.ts
@@ -14,7 +14,7 @@ import type {
     UserApiKeyRecord,
     UserApiKeyRepository,
 } from '@/infrastructure/db/types';
-import { withRetry } from '@/infrastructure/utils/withRetry';
+import { withRetry } from '@/shared/lib/withRetry';
 
 /**
  * Thrown when a stored ciphertext exists but cannot be decrypted with the
