+++ src/shared/api/isBot.ts
