+++ src/shared/lib/eastern.ts
