+++ src/shared/lib/marketSession.ts
